Angle (rad) and Length (microns, based on crd)
Combined the previous file angleinrad.m that calculates the angle between the filament and the horizontal, and a new variable, the length of the filament, assuming a straight line between the two points at the ends given by the raw coordinates crd (alternative to the arclen). Comparison between changes in length (with the length in microns and the arclen) and changes in angle may help determining which flips are actually part of a Jeffery orbit.