tau_r fixed for prolates with Nuris'formula
Nuris' thesis p. 47 gives the formula for the rotational diffusion time and coefficient for prolate bodies using Perrin's theory. Gives much more consistent results (in terms of magnitude). For a dynamic viscosity of 10^-3 Pa.s (apprx. by the dynamic viscosity of water), we get a tau_r of apprx. 30 sec.