ny -> NZ and nz -> NY
Change of coordinates from Andreas Zöttl's system (HORIZONTAL Hele-Shaw cell) to my experimental system (VERTICAL Hele-Shaw cell).
Consists basically in interverting the z and y axes. However, there may be more to it, in particular because in Andreas' case the z coordinate that was the unknown was estimated up to a sign, and becomes NY in our system, therefore the Jeffery constant C can only be estimated up to a sign as well (we're missing part of it).