Methods (1) and (2) to determine correlation time
Correlation time is estimated through:
--> (1) Under-estimate: first time the autocorr function crosses 0 and stays around (confidence interval) for at least a certain percentage of the total number of frames (default: 20%; default confidence bound: 0.05)
--> (2) Over-estimate: first time the autocorr function gets within a confidence interval and remains there for the entire rest of the experiment (default confidence bound: 0.1)
Plotting the two values on the graph, alongside with the smoothed out values of correlation time obtained with exponential fits (tau1 and tau2). The starting points of the fit can be modulated.

Essential data was modified to take into account the two different correlation times.